moved context to a seperate file
